1.1 update
Update to ensure the script runs more reliably by splitting the script into one section that starts on document-start and another that gives the UI window once the rest of the page is ready. Should now work in Firefox as well. 

Moved stuff around to keep static styling outside of UpdateUI so hopefully things are more consistent, and only the fontsize and theme colors are updated when updateUI is called. Have been tested that it works on OperaGX and Firefox.