RoR festival preset for highlight items
includes the default highlights (eggs, boss familiars, eliminate) as well as elemental holiday currency (currently all so i can just update the chest section for later fests) and the festival chests. To use, download the file and import it in the settings of the tool (highlight preset > import)! needs to be loaded manually so it's fully optional (and just an extra resource in case anyone wants it ..b)